service/pyproject.toml
[tool.poetry]
name = "head-count-hunter"
version = "1.0"
description = "head count hunter for burnout vision"
authors = ["AnuragAmbekar741 <anuragambekar1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
tortoise-orm = "^0.25.1"
uvicorn = "^0.37.0"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
aerich = "^0.9.1"
pydantic = "^2.12.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.3.0"
pydantic-settings = "^2.11.0"
python-dotenv = "^1.1.1"
google-auth = "^2.41.1"
pyjwt = "^2.10.1"
requests = "^2.32.5"
httpx = "^0.28.1"
langchain = "^1.0.7"
langchain-google-genai = "^3.0.3"
langchain-openai = "^1.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."